Add Input System

+++ Assets/Scripts/Object/OreConfig.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewOreConfig", menuName = "Config/Ore",order = 51)]
 public class OreConfig : ScriptableObject
 {
-    [SerializeField] private float _cost;
+    [SerializeField] private int _cost;
 
-    public float Cost => _cost;
+    public int Cost => _cost;
 }
